[COSMOS-4149] Citadel - Add file watcher for Root CA
Although it is not fully confirmed that the issue 4149 was caused
by the change of the root-cert.pem underneath, it was the only way
to reproduce the issue so far and thus the case described below
is treated as the main issue.

The Pilot-Agent is not able to generate certificate for the
application running in mTLS mode if the root certificate changes
while istiod and cert-manager are being down (due to connection
issues for example). After these components are back, the Citadel
clients uses the same connection with the usage of already invalid
certificates which makes it impossible to set a proper connection
and so the pilot-agent cannot retrieve the new mTLS certificates
so after 1 hour from the last update, the application becomes
unresponsive.

It is solved by adding a certificate watcher in the Citadel client,
which observes the changes on the root certificate. If the change
occurs, the Citadel will refresh the connection during the nearest
CSRSign call. However, if the certificate was deleted, the Citadel
will avoid restarting the connection in order to not break the
existing connection and then fail to establish a new connection if
there was an malfunction of the filesystem. In that scenario, the
Citadel client will wait for the certificate to show up again.