[COSMOS-4149] Citadel - Add tests for root cert watcher
This change adds a test confirming that the Citadel will
re-establish the connection if the root certificate changes.

To prepare this test, a new root certificate, certificate chain
and certificate key were generated using tools/certs makefile for
generating selfsigned certificate:
"make -f Makefile.selfsigned.mk new-cacerts"
Then the proper certificate files were moved to the testdata
directory.

These certs are being used as the replacement of previous
certificates. The scenario here is that both Citadel and the
service start with previous cert files and in the middle, the
service is shut down and the certs are being replaced with the new
certs. After placing new certs, the service is brought back and
we expect that the Citadel will be able to call CSRSign successfully
without the need for creating a new Citadel client.

This scenario is supposed to mimic a situation, where an root
istio ca is changed due to some malfunctions in the environment.
(more about this is described in previous commit and in the github
issue: https://github.com/istio/istio/issues/46494).

The test described above fails without the root cert watcher but
it succeeds now proving that the change works properly.